requirements.txt
llama-index-core>=0.10.0
llama-index-vector-stores-faiss>=0.1.0
llama-index-embeddings-huggingface>=0.1.0
llama-index-llms-ollama>=0.1.0
sentence-transformers>=2.2.2
python-frontmatter>=1.0.0
faiss-cpu>=1.7.4
rich>=13.4.2
requests>=2.31.0
